requirements.txt
transformers
torch
gradio
pillow
pytesseract
pdf2image